pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "cap-upskilling"
version = "0.1.0"
description = "CAP Upskilling Project"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
# FastAPI Backend Dependencies
fastapi = {version = "^0.119.0", extras = ["standard"]}
pydantic = "^2.11.9"
pydantic-settings = "^2.10.1"
httpx = "^0.28.1"
# Azure Dependencies
azure-identity = "^1.25.0"
azure-cognitiveservices-speech = "^1.43.0"
azure-search-documents = "^11.6.0"
# AI Dependencies
azure-ai-projects = "^1.0.0"
azure-ai-agents = ">=1.2.0b2"
opentelemetry-sdk = "^1.37.0"
azure-core-tracing-opentelemetry = "^1.0.0b12"
opentelemetry-api = "^1.37.0"
openai = "^1.109.0"
autogen-agentchat = "^0.7.4"
autogen-ext = { extras = ["mcp"], version = "^0.7.4" }
tiktoken = "^0.11.0"
agent-framework = {version = "^1.0.0b251007", allow-prereleases = true}
fhir-resources = "^8.1.0"
jinja2 = "^3.1.6"
orjson = "^3.11.3"
beautifulsoup4 = "^4.14.2"
scikit-learn = "^1.6.1"
# RAG Evaluation Dependencies
pandas = "^2.0.0"
numpy = "^2.2.6"
python-dotenv = "^1.1.0"
kaggle = "^1.6.0"

# Evaluation Dependencies
jiwer = "4.0.0"
rapidfuzz = "3.14.1"
azure-ai-evaluation = "^1.13.0"
azure-ai-ml ="^1.30.0"
mlflow = "^3.1.0"
azureml-mlflow ="1.60.0"

# Summarization Evaluation Dependencies
rouge-score = "^0.1.2"
nltk = "^3.9"
sentence-transformers = "^3.3.0"  # Clinical/medical embeddings support
scipy = "^1.14.0"  # For additional distance metrics
# PII Detection Dependencies
presidio-analyzer = "^2.2.360"
presidio-anonymizer = "^2.2.360"
datasets = "4.0.0"
 

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
mypy = "^1.8.0"
ipykernel = "^6.29.5"
python-dotenv = "^1.1.0"
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.1.1"

# Ruff configuration - linting and formatting
pytest-httpx = "^0.35.0"
poetry-plugin-dotenv = "^3.1.0"
[tool.ruff]
line-length = 240
target-version = "py313"
show-fixes = true
src = ["."]
exclude = ["src/frontend"]

[tool.ruff.lint]
select = [
    "B0",     # bugbear
    "E",      # default
    "F",      # default
    "UP",     # pyupgrade
    "RUF100", # valid noqa annotations
    "C90",    # mccabe complexity
    "N",      # pep8-naming
    "W",      # pycodestyle warnings
    "S",      # bandit security
    "T20",    # flake8-print
    "PIE",    # flake8-pie
    "I",      # isort import sorting
    "D",      # pydocstyle - docstring conventions
    "TID",    # flake8-tidy-imports - import conventions
]

ignore = [
    "D401",   # First line should be in imperative mood
    "S104",   # Possible binding to all interfaces
    "T201",   # Print statements allowed for CLI output
    "D102",   # All methods require a docstring - not required for simple/obvious methods
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "D"] # Allow asserts in tests, ignore docstring rules in tests
"**/__init__.py" = ["D104"] # Allow missing docstrings in __init__ files
"eval/**" = ["TID252"] # Allow relative imports in evaluation package
"eval/main.py" = ["E402"] # Cache paths must be set before imports
"eval/rag/main.py" = ["E402", "C901"] # Cache paths must be set before imports; complex metrics computation
"eval/eval_logger/main.py" = ["C901"] # Dynamic metrics logging with introspection
"eval/diarization/main.py" = ["C901"] # Complex bipartite graph connectivity logic for BER metric
"spikes/**" = ["TID252"] # Allow relative imports in experimental/spike code
# Healthcare/FHIR domain complexity exceptions - these functions handle complex medical data processing
"src/backend/service/clinical_profile/adapters/fhir_adapter.py" = ["C901"] # FHIR pagination logic
"src/backend/service/clinical_profile/service.py" = ["C901"] # Patient demographics extraction
"src/backend/service/clinical_profile/summarization/clinical_formatter.py" = ["C901"] # Medical screening logic
"src/backend/service/clinical_profile/summarization/pipeline.py" = ["C901"] # FHIR bundle normalization

[tool.ruff.lint.pydocstyle]
convention = "numpy"  # Use NumPy style docstrings

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
force-single-line = false
force-sort-within-sections = true
lines-between-types = 0  # Explicitly set to avoid warning when force-sort-within-sections is true
known-first-party = ["src", "eval", "tests"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.flake8-tidy-imports]
# Disallow all relative imports
ban-relative-imports = "all"

# Test configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "tests"]
addopts = [
    "--verbose",
    "--tb=short",
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
    #"--cov-fail-under=80"
]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

# Coverage configuration
[tool.coverage.run]
source = ["src/backend"]
omit = ["*/tests/*", "*/venv/*", "*/.venv/*", "src/frontend/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

# MyPy configuration
[tool.mypy]
python_version = "3.13"
show_error_codes = true
show_column_numbers = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = false # Set to true for stricter typing
no_implicit_optional = true
strict_optional = true
